RFC, TEST ONLY: add latency fairy to CQL server
An additional fiber is added to gather various metrics and
deduce possible concurrency limits.
Heavy assumptions/shortcuts:
 - requests are uniform, all use the same amount of resources
 - all requests are both reads and writes at the same time (yup, makes
   no sense)

This is just a very early RFC, but if we had more stats of requests
split by categories, e.g.:
 * reads
 * writes
 * cas reads
 * cas writes
 * (etc.)
... the estimated could be more sensible.

For specific scenarios, e.g. read-only workload of a single table
with specified schema and row size, or write-only workload of such
a table, results were very close to reality.

If we manage to create a good enough heuristics, it could later
be used to automatically set the concurrency limit, using only
the latency limit provided by the user - e.g. per session,
to allow OLTP/OLAP mixes, assuming that a single session is either
one or the other.