cql3: add preserving default values for per-sl timeouts
In order for per-service-level timeouts to work as expected,
a special value is reserved for internally marking the timeouts
as deleted.
With that, setting timeouts works as expected, e.g. statements
 ALTER SERVICE LEVEL sl WITH read_timeout = 5s;
 ALTER SERVICE LEVEL sl WITH write_timeout = 10s AND cas_timeout = 24h;
will result in setting read timeout to 5s, write timeout to 10s
and cas timeout to 24h. Previously, read_timeout value was erased
by the second statement due to having an implicit null value.