storage_proxy: disqualify read replicas likely to be dead
When a replica failed to send a successful response to the coordinator
for the time longer than the current request's remaining timeout,
it is now disqualified from taking part in the read operation.
The intention of this mechanism is to prevent a sharp increase
of background reads in a time window between a replica crashing
and this fact being detected by the failure detector
(which is several seconds).

Tests: unit(dev)
       manual(by crashing one of the replicas during a quorum
              read on a 3-node cluster and observing logs;
              including custom queries which use `USING TIMEOUT`)