qos: add a way of merging service level options
In order to combine multiple service level options coming from
multiple roles, a helper function is provided to merge two
of them. The semantics depend on each parameter, but for timeouts,
which are the only parameters at the time of writing this message,
the minimum value of the two is taken. That in particular means
that when service level A has read_timeout = 50ms, write_timeout = 10s
and no other timeouts, while service level B has read_timeout = 1s
and range_read_timeout = 5s, the resulting service level options
would set the timeouts to:
 - read timeout: 50ms
 - write timeout: 10ms
 - range_read_timeout: 5s